refactor: cleanup render method
